バーの表示調整

+++ c_step13/color_calc.py
@@ -62,7 +62,7 @@ def convert_height_to_byte(bar_height, box_height):
 
 def convert_byte_to_height(color, box_height):
     """色を、バーの縦幅ピクセルに変えます"""
-    return color/255*box_height
+    return int(color/255*box_height)
 
 
 def convert_3rates_to_3bytes(n3rates):
